# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Image metadata API with EXIF extraction and geolocation",
        "title": "Trekka API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@trekka.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Check if the API is running",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "status: ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/image": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve an image from Firebase Storage by filename",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Get an image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image filename",
                        "name": "fileName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to signed URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/images/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a paginated list of images with metadata from Firestore",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "List images",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1000,
                        "description": "Number of items to return (max 1000, default 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number (0-indexed, default 0)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of images",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/trekka-api_internal_models.ImageMetadata"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "trekka-api_internal_models.Coordinates": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "string"
                },
                "lng": {
                    "type": "string"
                }
            }
        },
        "trekka-api_internal_models.ImageMetadata": {
            "type": "object",
            "properties": {
                "contentType": {
                    "type": "string"
                },
                "coordinates": {
                    "$ref": "#/definitions/trekka-api_internal_models.Coordinates"
                },
                "createdAt": {
                    "description": "When record was created",
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "formattedDate": {
                    "description": "Format: \"Wednesday, 15 January 2025, 14:30\"",
                    "type": "string"
                },
                "geoLocation": {
                    "description": "Format: \"City, Country\"",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "resolution": {
                    "description": "Format: [width, height]",
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "storagePath": {
                    "type": "string"
                },
                "takenAt": {
                    "description": "Actual photo capture time from EXIF",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "When record was updated",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}